Initial Commit
Files are up to date. The .env file can be used to put the OPEN_API_KEY and the OPENAI_ORG_KEY. This file has to be added to the .gitignore file to not track the API keys. The overall application is functional and users can upload images and the chatbot will respond according to the detections made from YOLO (.pt file with the weights). LangChain needs to be integrated to read documents of different types.